-Fixed Negamax move searching
Issue: because the evaluation methods scale their score to the move passed as an argument, it makes a big difference how deep the search goes.  The best way to handle this would be to let them all use the traditional "more positive numbers signify a favorable position for white, and negative signifies a better position for black".

The engine is now making very logical choices for moves (for the most part).  In some cases I've seen it drop pieces, but I haven't found out for sure what causes this.  The next step is optimization so that the engine can look deeper without taking forever (2-3 ply search currently takes ~3 seconds)