Iterative Deepening added, Principal Variation move ordering complete
-After adding Principal variation to the move ordering consideration, the amount of nodes visited increased dramatically. After more research I learned this is because PV is ususally used with Iterative Deepening, (so you scope your favorable nodes in the previous iteration, not the previous move all together.  ID brought the amount of nodes visited for 5,4, and 3 ply searches very close to the optimal values.

Note: The engine is searching to depth 5 quite quickly now, but it seems to be dropping more pieces.  I think is is because of the lack of quiesent searching, which is next on my TODO list. After that the engine should be stable and I'll refactor some of the structure of the program (multithreading for GUI/passive analysis, so the game doesn't seem to freeze while the AI is thinking. This would allow for a progress bar to reflect the status of the AI's progress, which doesn't currently work without multithreading).

The issue of the BoardController trying to set pieces outside of the board bounds (-1 or 8) is persisting, and seems to happen mostly with pawns.  This could be causing the engine to miss moves because it mistakenly tries to put it on the wrong end square (which isn't legal) and then thinks the real move is illegal because BoardController throws an error if it isn't a legal operation.  This will be addressed after QS.