Checkmate miss bug fixed -Added check in Evaluate method to check for each side being checkmated, and to override the calculated value of the position if so.  This seems to have fixed the problem where an obvious checkmate is possible and the AI doesn't choose it.
TODO: Account for draws by 3 fold repition and stalemate.  I've seen the AI fall for these in a winning position in many games because it doesn't realize what is going on.

Also the AI seems to still be dropping pieces occasionally. It's hard to see what is causing that but I need to dig into that.