Null Move Search reduction fixed -Forgot that pvSearch assumes a valid set of child nodes, so those had to be cached and a new set generated of the opposite color since the previous children were of the opposite color as their parent. Since we're looking at a null move they needed to be the same color.  After the search the real children are restored
Engine plays pretty quickly with this in place. It almost never looks farther than the 1st or 2nd branch, but hopefully that's a sign of good move ordering. Still getting some "king not found" exceptions in the end game, and the game crashed when I tried to bump it up to depth 6, but once that is working I think the engine will be quite strong.  It got Chess Titans to a rook + minor piece endgame and was 1 pawn up with black when it crashed on Level 7 at depth 5, so I'm ambitous about what it can do at higher depths now.