Performance tuning, PV search fix -PV Search now caches the child nodes themselves, not a list of them. This keeps it from replacing the parent node's children with the ones of the same color -Piece type is now a char instead of a string -No more local legalMove lists for Knights, Rooks, and Bishops, since the arrayList addAll operation was taking up a lot of time.
Issues: Game is super slowing down after a few moves now.  This doesn't seem to be related to the PV Search additions, but it is rendering the game pretty useless right now. Defintely a priority to track down what is causing it.

Next: Controller.RemovePieceFromList seems to be the next most expensive method in the profiler.  Might construe a special kind of static array for this, since there are only ever 16 pieces on a side. That could get constant look up time instead of looping over an arrayList