Added refactoring and debugging to try and track down the fool's mate bug
Notes from analysis over the weekend:

Board array still thinks queen on 7,3. She moved from 55 to 22.

It's unclear how she got to 5,5. Never moved there. Some data isn't getting sanitized in make unmake process

Board thinks queen is on 5,5.
makes a move from 5,5 to 2,2
end of the story: queen is not on the square it was thought she moved to.  en passent assumes it means  en pass okay because it checks to see if there is no piece on the end square of the previous move.